Time: 4 ms (44.03%), Space: 5.8 MB (82.70%) - LeetHub

+++ 1359-count-all-valid-pickup-and-delivery-options/1359-count-all-valid-pickup-and-delivery-options.cpp
@@ -4,7 +4,7 @@ public:
         
         if(n == 1) return 1;
         
-        long long int result = 1;
+        long int result = 1;
         
         for(int i=1 ; i<=n ; i++) {
             // Here, i for factorial and (2*i-1) for series
